### 2019-10-31 #1851 @whedon
[PRE REVIEW]: workloopR: Analysis of work loops and other data from muscle physiology experiments in R
**Submitting author:** @vbaliga (<a href="http://orcid.org/0000-0002-9367-8974">Vikram B. Baliga</a>)
**Repository:** <a href="https://github.com/ropensci/workloopR/" target ="_blank">https://github.com/ropensci/workloopR/</a>
**Version:** v1.1.1
