### 2019-10-31 #1851 @Kevin-Mattheus-Moerman
Note: this package has been reviewed by rOpenSci: https://github.com/ropensci/software-review/issues/326
