### 2019-10-31 #1851 @vbaliga
Hi @Kevin-Mattheus-Moerman - thank you for your help. I can confirm that this package has been reviewed by rOpenSci: https://github.com/ropensci/software-review/issues/326, @lmullen was the editor

I have just taken a look a the proof that whedon provided in https://github.com/openjournals/joss-reviews/issues/1851#issuecomment-548467135 . I am generally happy with how things look, but there is one mistake that I would like to fix: I neglected to list our sources of financial support in our Acknowledgments.

I'd like to append the following information to the Acknowledgments: This work was funded by the U.S. Air Force Office of Scientific Research (under grant no. FA9550-16-1-0182, titled "Avian-inspired multifunctional morphing vehicles" monitored by B. L. Lee) and also by the Natural Sciences and Engineering Research Council of Canada (NSERC, RGPIN-2016-05381).

Do you have advice on the most efficient way to proceed? Should I update `paper.md`  in https://github.com/ropensci/workloopR  and create a new release to trigger Zenodo? Or is there another preferred way? Thanks!
