### 2019-10-31 #1851 @danielskatz
Please go ahead and make the paper fix in the .md file.  If you want to change the version, that's fine - If you want to make a new archive of the software, that's fine too.
