### 2019-11-01 90ee92b
fix acknowledgments in JOSS submission paper
add our funding sources to the paper. also small fixes to language & typos, and added Issues info to our README
