+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: workloopR
 Type: Package
 Title: Analysis of Work Loops and Other Data from Muscle Physiology Experiments
-Version: 1.1.0
+Version: 1.1.1
 Authors@R: 
     c(person(given = "Vikram B.",
              family = "Baliga", 
