### 2019-11-01 #1851 @vbaliga
Hi @danielskatz - thank you for your advice. I went ahead and made the fix to the paper and [bumped the version](https://github.com/ropensci/workloopR/releases/tag/v1.1.1) to 1.1.1 . Our Zenodo archive ([https://doi.org/10.5281/zenodo.3523384](https://doi.org/10.5281/zenodo.3523384)) has also been updated, following the new version.

I think our JOSS submission is now ready. Please let me know if there is any further action I should take. Thanks again
