### 2019-11-01 #1851 @vbaliga
> [ 👉 Check article proof 📄 👈 ](https://github.com/openjournals/joss-papers/blob/joss.01851/joss.01851/10.21105.joss.01851.pdf)

The proof generated here looks good to me
