### 2019-11-01 #1856 @danielskatz
Note: this package has been reviewed by rOpenSci: https://github.com/ropensci/software-review/issues/326
