### 2019-11-01 #1856 @danielskatz
👋 @vbaliga - in the paper, can you please add commas after "e.g."s  so that they look like, e.g., this.)
Also, please fix the cases in the first reference - the first word doesn't need to all CAPS...
