### 2019-11-01 0fc4be0
fix case and commas in JOSS paper
add commas after "e.g." and fix the title of Ahn 2012 so the first word is not the only one in all caps
