+++ paper.bib
@@ -13,7 +13,7 @@
 
 @article{Ahn:2012,
     author = {Ahn, A. N.},
-    title = {HOW MUSCLES FUNCTION {\textendash} THE WORK LOOP TECHNIQUE},
+    title = {How muscles function {\textendash} the work loop technique},
     volume = {215},
     number = {7},
     pages = {1051--1052},
