+++ README.md
@@ -2,8 +2,8 @@ workloopR <img src="https://github.com/ropensci/workloopR/raw/master/images/work
 =======================================================================================
 
   <!-- badges: start -->
-[![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)  [![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/ropensci/workloopR?branch=master&svg=true)](https://ci.appveyor.com/project/ropensci/workloopR)  [![Travis build status](https://travis-ci.org/ropensci/workloopR.svg?branch=master)](https://travis-ci.org/ropensci/workloopR)
-[![](https://badges.ropensci.org/326_status.svg)](https://github.com/ropensci/software-review/issues/326)
+[![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)  [![Travis build status](https://travis-ci.org/ropensci/workloopR.svg?branch=master)](https://travis-ci.org/ropensci/workloopR)
+[![](https://badges.ropensci.org/326_status.svg)](https://github.com/ropensci/software-review/issues/326)  [![status](https://joss.theoj.org/papers/f0eca1d99dde4e02b414e65271cc24ca/status.svg)](https://joss.theoj.org/papers/f0eca1d99dde4e02b414e65271cc24ca)
   <!-- badges: end -->
 
 Analysis of work loops and other data from muscle physiology experiments in R
@@ -79,7 +79,7 @@ wl_analyzed_simple
 
 *For an overview, please also see our ["Introduction to workloopR"](https://docs.ropensci.org//workloopR/articles/Introduction-to-workloopR.html) vignette*
 
-**Data import**: Importing data creates objects of class `muscle_stim`, which we designed to essentially behave as `data.frame`s but with unique properties that work nicely with ``workloopR``'s core functions. Data that are stored in .ddf format (e.g. generated by Aurora Scientific's Dynamic Muscle Control and Analysis Software) are easily imported. Other file formats are welcome, but need to be constructed into `muscle_stim` objects by the user; please see the vignette ["Importing data from non .ddf sources"](https://docs.ropensci.org/workloopR/articles/non-ddf-sources.html).
+**Data import**: Importing data creates objects of class `muscle_stim`, which we designed to essentially behave as `data.frame`s but with unique properties that work nicely with ``workloopR``'s core functions. Data that are stored in .ddf format (e.g., generated by Aurora Scientific's Dynamic Muscle Control and Analysis Software) are easily imported. Other file formats are welcome, but need to be constructed into `muscle_stim` objects by the user; please see the vignette ["Importing data from non .ddf sources"](https://docs.ropensci.org/workloopR/articles/non-ddf-sources.html).
 
 - Example:
 
@@ -129,7 +129,7 @@ analyze_twitch <- isometric_timing(twitch_dat,
 ```
 
 
-**Batch processing**: We also include functions for batch processing files (e.g. multiple files from a common experiment). These functions allow for the import, cycle selection, gear ratio correction, and ultimately work & power computation for all work loop trial files within a specified directory.  This also allows users to correct for potential degradation of the muscle (according to power & work output) over the course of the experiment. See the ["Batch processing"](https://docs.ropensci.org/articles/batch-processing.html) vignette
[…]
