### 2019-11-01 #1856 @vbaliga
Hi @danielskatz - thanks for catching these. I pushed a new commit (https://github.com/ropensci/workloopR/commit/0fc4be098671cfd0fb2d3ab7f72480309cbc767e) that takes care of both the comma issue and the CAPS issue. I did not create a new release, so the one I made yesterday ([v1.1.1](https://github.com/ropensci/workloopR/releases/tag/v1.1.1)) is still the latest one in our Zenodo archive (https://doi.org/10.5281/zenodo.3523384), too. Please let me know if there's anything more I should do & thanks again for all your help
