### 2019-11-01 #1856 @danielskatz
@vbaliga - one more small change is needed - the first reference title should have a "-" in the middle: "How muscles function - the work loop technique"
