### 2019-11-01 6323524
fix endash in Ahn 2012
per https://github.com/openjournals/joss-reviews/issues/1856#issuecomment-548812646
