+++ paper.bib
@@ -13,7 +13,7 @@
 
 @article{Ahn:2012,
     author = {Ahn, A. N.},
-    title = {How muscles function {\textendash} the work loop technique},
+    title = {How muscles function - the work loop technique},
     volume = {215},
     number = {7},
     pages = {1051--1052},
