### 2019-11-01 #1856 @vbaliga
@danielskatz - OK no problem. I just pushed a commit (https://github.com/ropensci/workloopR/commit/632352453351bc86991242aa9cf79195341c2ba1) that replaced the dash. I think that should take care of the issue.
