### 2019-11-01 #1856 @vbaliga
Hi @danielskatz,
Thank you sincerely for all your help! @shreeramsenthi and I both appreciate your patience in getting us through the extra steps and we are very happy to now be published in JOSS.
Best regards,
Vikram Baliga
